fixed entity type, was casting down to size_t

+++ MoonRuntime/Source/ECS/Entity.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 // A simple type alias
-using Entity = unsigned int;
+using Entity = size_t;
 
 // Used to define the size of arrays later on
 const Entity MAX_ENTITIES = 10000;
